AnalysisService/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic>=1.8.0
langchain>=0.1.0
langchain-together>=0.0.1
langchain-openai>=0.0.1
aperturedb==0.4.40
weave>=0.1.0
pyyaml>=5.4.1
openai
langgraph
langchain-community
sentence-transformers
fastapi-cache2>=0.1.9
